+++ src/Laravel/TelegramServiceProvider.php
@@ -8,6 +8,7 @@ use Illuminate\Support\ServiceProvider;
 use Illuminate\Contracts\Container\Container as Application;
 use Laravel\Lumen\Application as LumenApplication;
 use Illuminate\Foundation\Application as LaravelApplication;
+use Telegram\Bot\Laravel\Artisan\WebhookCommand;
 
 /**
  * Class TelegramServiceProvider.
@@ -82,7 +83,7 @@ class TelegramServiceProvider extends ServiceProvider
         });
 
         $this->app->singleton('telegram.bot.commands.webhook', function () {
-            return new WebHookArtisanCommand();
+            return new WebhookCommand();
         });
 
         $app->alias('telegram.bot', Api::class);
